Implement a state saver that supports the Bundle.
