+++ sample/app-android/build.gradle.kts
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2020-2021 Dmitriy Gorbunov (dmitriy.goto@gmail.com)
+ * Copyright (c) 2020-2022 Dmitriy Gorbunov (dmitriy.goto@gmail.com)
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -82,5 +82,6 @@ dependencies {
     implementation(Libs.composeActivity)
     implementation(Libs.androidxWindow)
     implementation(compose.preview)
+    implementation(Libs.bundlizer)
     implementation(project(":sample:app-common-compose-ui"))
 }
