+++ sample/app-common/src/commonMain/kotlin/me/dmdev/premo/sample/serialization/Serializers.kt
@@ -1,0 +1,91 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2020-2022 Dmitriy Gorbunov (dmitriy.goto@gmail.com)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package me.dmdev.premo.sample.serialization
+
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.modules.SerializersModule
+import me.dmdev.premo.PmDescription
+import me.dmdev.premo.sample.CounterPm
+import me.dmdev.premo.sample.MainPm
+import me.dmdev.premo.sample.SamplesPm
+import me.dmdev.premo.sample.bottom_navigation.BottomNavigationPm
+import me.dmdev.premo.sample.bottom_navigation.TabItemPm
+import me.dmdev.premo.sample.bottom_navigation.TabPm
+import me.dmdev.premo.sample.stack_navigation.SimpleScreenPm
+import me.dmdev.premo.sample.stack_navigation.StackNavigationPm
+
+object Serializers {
+    val module = SerializersModule {
+        polymorphic(
+            PmDescription::class,
+            MainPm.Description::class,
+            MainPm.Description.serializer()
+        )
+        polymorphic(
+            PmDescription::class,
+            MainPm.Description::class,
+            MainPm.Description.serializer()
+        )
+        polymorphic(
+            PmDescription::class,
+            SamplesPm.Description::class,
+            SamplesPm.Description.serializer()
+        )
+        polymorphic(
+            PmDescription::class,
+            CounterPm.Description::class,
+            CounterPm.Description.serializer()
+        )
+        polymorphic(
+            PmDescription::class,
+            StackNavigationPm.Description::class,
+            StackNavigationPm.Description.serializer()
+        )
+        polymorphic(
+            PmDescription::class,
+            SimpleScreenPm.Description::class,
+            SimpleScreenPm.Description.serializer()
+        )
+        polymorphic(
+            PmDescription::class,
+            BottomNavigationPm.Description::class,
+            BottomNavigationPm.Description.serializer()
+        )
+        polymorphic(
+            PmDescription::class,
+            TabPm.Description::class,
+            TabPm.Description.serializer()
+        )
+        polymorphic(
+            PmDescription::class,
+            TabItemPm.Description::class,
+            TabItemPm.Description.serializer()
+        )
+    }
+
+    val json =  Json {
+        serializersModule = module
+    }
+}
